feat(Application): create printer class feat(Application): create Constants Package for Messages and Constants Values
